HUGE REFACTORING TO GET LOGGING TO WORK. MY GOD.
