+++ imports/api/Fhir/fhirMethods.js
@@ -3,260 +3,9 @@
  * these methods will take the results from the query to the server and transform it for the front end to utilize 
  * when creating graphs off of patient medical data.
  */
-import { fhirClient } from './fhirclient.js';
+
 import { Meteor } from 'meteor/meteor';
-import { logger } from '../Logging/logger-config.js';
 import { LOINC_MAPPING } from '../Loinc/loincConstants.js';
-
-// takes only the useful information about the given observation
-function transformObservationInformation(observationResource) 
-{
-    let valueQuantities;
-
-    if (!observationResource || observationResource.resourceType !== "Observation")
-        return null;
-
-    if (observationResource.component && Array.isArray(observationResource.component)) {
-        valueQuantities = observationResource.component.map((subObservation) => {
-            return {
-                loincCode: subObservation.code.coding[0].code,
-                loincText: subObservation.code.coding[0].display,
-                valueQuantity: subObservation.valueQuantity
-            }
-        })
-    }
-
-    return {
-        loincText: observationResource.code.text,
-        loincCode: observationResource.code.coding[0].code,
-        dateIssued: observationResource.issued,
-        valueQuantities: observationResource.valueQuantity ? [observationResource.valueQuantity] : valueQuantities
-    }
-
-}
-
-async function transformDiagonosticReportInformation(diagnosticsReportResource) {
-    if(!diagnosticsReportResource || diagnosticsReportResource.resourceType !== "DiagnosticReport")
-        return null;
-
-    console.log(diagnosticsReportResource)
-    let resource = {
-        loincCode: diagnosticsReportResource.code.coding[0].code,
-        loincText: diagnosticsReportResource.code.coding[0].display,
-        dateIssued: diagnosticsReportResource.issued,
-        observations: diagnosticsReportResource.result
-    };
-
-    //map referenced observations to actual resources from fhir server
-    try {
-        resource.observations = await Promise.all(
-            resource.observations.map(async (ref) => {
-                const observationID = ref.reference.split("/")[1]; //Observation/1
-                const observationResource = await getPatientObservation(observationID);
-                return transformObservationInformation(observationResource);
-            }));
-    
-        return resource;
-    } 
-    catch (error) {
-        console.error(error.message);
-        throw error;
-    }
-    
-}
-function getPatientAge(birthdate) {
-    const dateNow = Date.now();
-    const patientBirthDateMs = Date.parse(birthdate);
-
-    const ageDate = new Date(dateNow - patientBirthDateMs);
-    return Math.abs(ageDate.getUTCFullYear() - 1970);
-}
-
-//returns the full FHIR patient record of the specified patientID
-async function getPatientRecordByID(patientID) {
-    let response;
-
-    try {
-        response = await fhirClient.read({
-            resourceType: "Patient",
-            id: patientID
-        });
-
-        if(response.birthDate)
-            response["age"] = getPatientAge(response.birthDate)
-        else 
-            response["age"] = "unknown";
-
-        return response;
-    } 
-    catch (error) {
-        console.log(error.message);
-        throw error;
-    }
-}
-
-/**
-*   Given required user information: given name, family name, phonenumber, dob. 
-*   This function will search through the fhir server to find all patient records 
-*   with all matching information. If no record is found, -1 is returned. If multiple
-    records are found, a list of IDs are returned. If only one record is found,
-*   the id of that record is returned.
-*/  
-async function findPatientByInfo(patientInformation) {
-
-    let matchedPatients = [];
-
-    const {patientGivenName, patientFamilyName, patientPhoneNumber, patientDOB} = patientInformation;
-
-    logger.info(patientInformation, "searching for patient with matching information in fhir server.")
-
-    try {
-        const searchResponse = await fhirClient.search({
-            resourceType: "Patient",
-            searchParams: {
-                given: patientGivenName, 
-                family: patientFamilyName,
-                birthdate: patientDOB,
-                telecom: patientPhoneNumber
-            }
-        });
-
-        if(!searchResponse || !searchResponse.entry || searchResponse.total === 0) {
-            logger.warn("no patient found");
-            return -1;
-        }
-    
-        if(searchResponse.total === 1) {
-            logger.info(`Patient found with fhirID ${searchResponse.entry[0].resource.id}`);
-            return parseInt(searchResponse.entry[0].resource.id);
-        }
-
-        for (const patient of searchResponse.entry) {
-            matchedPatients.push(parseInt(patient.resource.id));
-        }
-
-        if (matchedPatients.length > 1)
-            logger.warn("Multiple patients found matching information.")
-
-        return matchedPatients;
-    }
-    catch (error) {
-        logger.error(error.message);
-        throw error;
-    }
-}
-
-async function getPatientObservation(observationID) {
-    let response;
-    try {
-        response = await fhirClient.read({
-            resourceType: "Observation",
-            id: observationID
-        })
-        logger.info(`Getting observation # ${observationID}`)
-
-        return response;
-    }
-    catch (error) {
-        throw error;
-    }
-    
-}
-
-async function getPatientHealthMetrics(loincCode, patientID, pageNumber, count) {
-    let response;
-    let metrics = [];
-    let offset = 0;
-    
-    try {
-        logger.info({loincCode, patientID, pageNumber, count}, `Getting Health Metrics for Patient/${patientID}`);
-
-        if (!(loincCode && patientID && pageNumber && count)) {
-            logger.warn("Required health metric arguments were not present.");
-            return metrics;
-        }
-
-        //join loincCodes in a string seperated by "," and retrieve observations with the loinc Codes.
-        if (Array.isArray(loincCode)) {
-            loincCode = loincCode.join(",");
-        }
-        
-
-        if (pageNumber <= 0 || count <= 0) {
-            logger.warn("Page number and count out of bounds.");
-            return metrics;
-        }
-
-
-        offset = (pageNumber - 1) * count;
-
-        response = await fhirClient.search({
-            resourceType: "Observation",
-            searchParams: {
-                code: loincCode, 
-                subject: patientID,
-                _count: count,
-                _offset: offset,
-
-            },
-        });
-
-        if (!response || response.total === 0 || !response.entry) {
-            return metrics;
-        }
-        for(let entry of response.entry){
-            metrics.push(transformObservationInformation(entry.resource));
-        }
-
-        return metrics;
-
-    } 
-    catch (error) {
-        throw error;
-    }
-}
-
-/**
- * a function that pulls a given patient's diagnosticsreports from the fhir server and pulls the refrencing observations
- * so that the results can be displayed visually in the recent labs section of the client dashboard
- * @param {*} patientID 
- */
-async function getRecentPatientLabs(patientID, pageNumber, count) {
-    let labs = [];
-    let offset = (pageNumber - 1) * count;
-    try {
-        logger.info({patientID: patientID, pageNumber: pageNumber, count: count}, `retrieving patient labs for Patient/${patientID}`)
-        let searchResponse;
-
-        if (count <= 0 || pageNumber <= 0) {
-            return labs;
-        }
-
-        searchResponse = await fhirClient.search({
-            resourceType: "DiagnosticReport",
-            searchParams: {
-                subject: patientID,
-                category: "LAB",
-                _sort: "-date",
-                _offset: offset,
-                _count: count
-            }
-        });
-
-        if (!searchResponse || searchResponse.total === 0 || !searchResponse.entry)
-            return labs;
-
-        for (let entry of searchResponse.entry) {
-            labs.push(await transformDiagonosticReportInformation(entry.resource))
-        }
-
-        return labs;
-    }
-    catch (error) {
-        throw error;
-    }
-
-}
 
 Meteor.methods({
     /**
@@ -269,12 +18,15 @@ Meteor.methods({
      */
     async "patient.getHealthMetrics"(loincCode, patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(loincCode, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if(!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(loincCode, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
     /**
      * Retrieves a full FHIR patient record by ID.
@@ -283,181 +35,230 @@ Meteor.methods({
      */
     async "patient.getRecordByID"(patientID) {
         this.unblock();
-        try {
-            return await getPatientRecordByID(patientID);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientRecordByID } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientRecordByID(patientID);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getWeightMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_WEIGHT, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_WEIGHT, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getHeightMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_HEIGHT, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if(!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_HEIGHT, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getHeartRateMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_HEART_RATE, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if(!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_HEART_RATE, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getBMIMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_BMI, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_BMI, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+       
     },
     
     async "patient.getBloodPressureMetrics"(patientID, pageNumber = 1, count = 100) {
+
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_BLOOD_PRESSURE, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_BLOOD_PRESSURE, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getBodyTempMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_TEMP, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_TEMP, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }    
         }
     },
 
     async "patient.getBodyOxygenSaturationMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.BODY_OXYGEN_SATURATION, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.BODY_OXYGEN_SATURATION, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getHemoglobinHGBMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.HEMOGLOBIN_HGB, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.HEMOGLOBIN_HGB, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getHemoglobinA1CMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.HEMOGLOBIN_A1C, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.HEMOGLOBIN_A1C, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getErythrocyteSedimentationRateMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.ERYTHROCYTE_SEDIMENTATION_RT, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.ERYTHROCYTE_SEDIMENTATION_RT, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getGlucoseSerumPlasmaMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.GLUCOSE_SERUM_PLASMA, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.GLUCOSE_SERUM_PLASMA, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getPotassiumSerumPlasmaMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.POTASSIUM_SERUM_PLASMA, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.POTASSIUM_SERUM_PLASMA, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
+        
     },
 
     async "patient.getCholesterolTotalMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.CHOLESTEROL_TOTAL, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.CHOLESTEROL_TOTAL, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
 
     async "patient.getLowDensityLipoproteinMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.LOW_DENS_LIPOPROTEIN, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.LOW_DENS_LIPOPROTEIN, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
 
     async "patient.getHighDensityLipoproteinMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.HIGH_DENS_LIPOPROTEIN, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.HIGH_DENS_LIPOPROTEIN, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
 
     async "patient.getUreaNitrogenBUNMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.UREA_NITROGEN_BUN, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.UREA_NITROGEN_BUN, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
 
     async "patient.getCreatinineMetrics"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getPatientHealthMetrics(LOINC_MAPPING.CREATININE, patientID, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getPatientHealthMetrics } = await import("./Server/FhirUtils.js");
+            try {
+                return await getPatientHealthMetrics(LOINC_MAPPING.CREATININE, patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
     /**
@@ -469,11 +270,13 @@ Meteor.methods({
      */
     async "patient.getRecentLabs"(patientID, pageNumber = 1, count = 100) {
         this.unblock();
-        try {
-            return await getRecentPatientLabs(patientID, labReturnLimit, pageNumber, count);
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { getRecentPatientLabs } = await import("./Server/FhirUtils.js");
+            try {
+                return await getRecentPatientLabs(patientID, pageNumber, count);
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
     /**
@@ -487,13 +290,13 @@ Meteor.methods({
      */
     async "patient.findByInfo"({ patientGivenName, patientFamilyName, patientPhoneNumber, patientDOB }) {
         this.unblock();
-        try {
-            return await findPatientByInfo({ patientGivenName, patientFamilyName, patientPhoneNumber, patientDOB });
-        } catch (error) {
-            logger.error(error, "FHIR-Server-Error");
-            throw new Meteor.Error("FHIR-Server-Error", error.message);
+        if (!this.isSimulation) {
+            let { findPatientByInfo } = await import("./Server/FhirUtils.js");
+            try {
+                return await findPatientByInfo({ patientGivenName, patientFamilyName, patientPhoneNumber, patientDOB });
+            } catch (error) {
+                throw new Meteor.Error("FHIR-Server-Error", error.message);
+            }
         }
     },
 });
-
-module.exports = { transformObservationInformation };
