+++ imports/startup/Server/initRoles.js
@@ -1,5 +1,5 @@
 import { Roles } from 'meteor/alanning:roles';
-import { UserRoles } from "../api/User/userRoles.js";
+import { UserRoles } from "../../api/User/userRoles.js";
 
 Meteor.startup(async () => {
