+++ server/main.js
@@ -2,7 +2,8 @@ import "../imports/api/Fhir/fhirMethods.js";
 import "../imports/api/Loinc/loincConstants.js";
 import "../imports/api/User/userMethods.js";
 import "../imports/api/Loinc/lioncMethods.js";
-import "../imports/startup/accounts-config.js";
-import "../imports/startup/initRoles.js";
-import "../imports/startup/initUsers.js";
-import "../imports/api/Logging/logger-config.js";
+import "../imports/api/Logging/Server/logger-config.js";
+import "../imports/startup/Server/accounts-config.js";
+import "../imports/startup/Server/initRoles.js";
+import "../imports/startup/Server/initUsers.js";
+
